+++ CHANGES.rst
@@ -30,6 +30,9 @@ New Features
   - Distance calculations in cosmologies with massless (or no) neutrinos
     are now > 10x faster for the supplied cosmologies. [#3767]
 
+  - Distances with massive neutrinos are 3-6x faster for the most
+    common cases of 1-3 massive neutrino species. [#3813]
+
   - ``FLRW._tfunc`` and ``FLRW._xfunc`` are marked as deprecated.  Users
     should use the new public interfaces ``FLRW.lookback_time_integrand`` 
     and ``FLRW.abs_distance_integrand`` instead. [#3767]